修复语法错误

+++ src/main/kotlin/JHorseRacing.kt
@@ -195,12 +195,11 @@ object JHorseRacing : KotlinPlugin(
                         // 胜场计数累积
                         stat.winCount += 1
                         // 收益积分
-                        (bet.score * if (bet.score == score)
-                            JHRPluginConfig.allinOdds // 若allin，则使用allin赔率
-                        else
-                            // 否则根据下注金额与余额的比值计算赔率
-                            (bet.score.toDouble() / 1000.0 * JHRPluginConfig.maxOdds)
-                                .coerceIn(JHRPluginConfig.minOdds, JHRPluginConfig.maxOdds)
+                        (bet.score *
+                            if (bet.score == score) JHRPluginConfig.allinOdds
+                            else (bet.score.toDouble() / 1000.0 * JHRPluginConfig.maxOdds)
+                                     .coerceIn(JHRPluginConfig.minOdds, JHRPluginConfig.maxOdds)
+                        ).toInt()
                     } else {
                         // 亏损积分
                         -bet.score
